pyproject.toml
[project]
name = "lunch-menu-refactoring"
version = "0.1.1"
description = "It's about menu-eda"
authors = [
    {name = "jacob-53", email = "jacob8753@gmail.com"},
]
dependencies = ["pandas>=2.2.3", "streamlit>=1.42.0", "matplotlib>=3.10.0", "psycopg>=3.2.4", "python-dotenv>=1.0.1"]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "BSD"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
